Add support for user-supplied palette + indexed pixel input
Quantizing (palette computation + remapping RGB to indexed) is a slow
process.  If the gif encoder inputs already happen to be indexed
(which it often is when say encoding pixel art into gif anims), this
whole process can be skipped.

This patch allows specifying a palette once on GIFEncoder creation and
makes it possible to pass indexed inputs to GIFEncoder.addFrame.

This is somewhat of a prototype.  Can clean this up if it looks like
this has a chance to be merged.